final_project/requirements.txt
torch
torchvision
Pillow
matplotlib
tqdm
dataclasses
openai-clip
numpy